fix: ignore wrappers (#119)
Hello,
In NixOS it is impossible to use the wrappers from mvn and gradle and we
have to always use the binaries which are installed via the derivations.

It looks like the `ignore_wrappers` option was ignored completely, and
there was also a small bug when setting the options (it seems to me the
API for `vim.tbl_extend` has changed, and it now returns a table instead
of modifying the table in-place). I fixed both issues in this MR.
Please tell me if there's something more I can do :)