Implement more robust runtime collection and compilation
On top of the existing runtime this improvement also tries to check the
maven or gradle (todo) to collect as much information about the runtime
of the project, if that is not possible, the user might be prompted to
enter the home directory of the runtime, the user might also configure
the runtimes statically in the plugin's configuration.

Furthermore dependencies on other plugins such as nvim-jdtls is removed,
and the changes rely only on native the nvim lsp or coc.nvim.