Log error when modulepaths element is set.
For a project using the Java Platform Module System, the modulepaths and
classpaths elements are populated by eclipse jdtls for the
java.project.getClasspaths command. This is a problem because
neotest-java does not use the modulepaths element. A second problem
exists when these elements are populated for a maven project. For a
maven project the target/test-classes element is omitted from the
generated paths. A simple work around to insure successful test
execution is to temporarily remove/rename the module-info.java file.