fix: reinstate nearest functionality #203
The 'Run Nearest Test' and 'Debug Nearest Test' options do operate as
expected, and run/debug all the tests in a class instead of just the
nearest. This change was a work around introduced because of a bug in
 a prior version of junit-platform-console-standalone.

As of junit-platform-console-standalone version 1.10.1, invocation with
method option '-m' is correct so this patch once again reinstates the
'nearest' functionality.