feat: create positions with FQN as id (#209)
This PR address the problem of running parameterized test invoking the
correct parameter for the JUnit5 command.

This also fixes the problem of running just the methods that are
actually selected. (#203)

The correct FQN for the method in this code snippet:

```java
    package com.example;

    import org.junit.jupiter.params.ParameterizedTest;
    import java.util.List;
    import java.util.Map;

    class SomeParameterizedTest {

      @ParameterizedTest
      void testWithParameters(String name) {}
    }
```

Before: `com.example.SomeParameterizedTest#testWithParameters` ❌ 

Good:
`com.example.SomeParameterizedTest#testWithParameters(java.lang.String)`
✅